Generate count and discriminator nodes
This PR extends the spec generator to cover two more node categories \u2014 `count` and `discriminator` \u2014 landing six generated structs and two union enums under `codama-nodes/src/generated/`. Hand-written `<category>/<node>.rs` files shrink to constructors and `#[cfg(test)] mod tests`, mirroring the PR #1 treatment of link nodes. Integer fields backed by spec `integer/u64` (the `value`, `offset`, and `size` attributes) become `u64` for determinism; the change propagates through the four constructors, the `ArrayTypeNode`/`SetTypeNode`/`MapTypeNode` `.fixed()` helpers, the `discriminator_directive`'s `SetOnce`s, and the `identify_field_types_visitor`'s array-length parsing, all of which now flow `u64` end to end. The generator gains a small heuristic that derives `Copy` when every attribute is a scalar (integer / float / boolean / enumeration) and `Default` when the struct is empty, eliminating per-node configuration. `getEmittableUnions` now requires a `registered<Name>` sibling, which keeps the category-main unions and skips inline / synthetic ones. A separately-fixed bug: optional `docs` attributes were being wrapped in `Option<Docs>`; they are now emitted as bare `Docs` regardless of the spec's `optional` flag, matching the established convention. `pdaSeed` is intentionally left out of this PR \u2014 its `constantPdaSeedNode.value` references a nested inline union (`constantPdaSeedValue`) that requires the non-flattening union machinery planned for a later PR.

+++ codama-attributes/src/codama_directives/discriminator_directive.rs
@@ -35,8 +35,8 @@ impl DiscriminatorDirective {
         let mut encoding =
             SetOnce::<BytesEncoding>::new("encoding").initial_value(BytesEncoding::Base16);
         let mut field = SetOnce::<CamelCaseString>::new("field");
-        let mut offset = SetOnce::<usize>::new("offset").initial_value(0);
-        let mut size = SetOnce::<usize>::new("size");
+        let mut offset = SetOnce::<u64>::new("offset").initial_value(0);
+        let mut size = SetOnce::<u64>::new("size");
         pl.each(|ref meta| match meta.path_str().as_str() {
             "bytes" => {
                 if kind != DiscriminatorKind::Constant {
