+++ codama-korok-visitors/src/identify_field_types_visitor.rs
@@ -97,7 +97,7 @@ pub fn get_type_node(ty: &syn::Type) -> Option<TypeNode> {
             }
         }
         syn::Type::Array(syn::TypeArray { elem, len, .. }) => {
-            let Ok(size) = len.as_unsigned_integer::<usize>() else {
+            let Ok(size) = len.as_unsigned_integer::<u64>() else {
                 return None;
             };
             get_type_node(elem)
